reorganize code, clean up default ordering logic
